Fix compilation warning on 64bit arch in datatore_postgres and sqlite. Fix completion bug when given word is longer than expected one. Fix conntrack bug that corrupted the linked list of target and helper priv when closing them. Fix small memory leak in target_rtp. Some more hardening of match_rtp.

+++ src/datastore_postgres.c
@@ -408,7 +408,7 @@ static int datastore_dataset_read_postgres(struct datastore *d, struct dataset *
 						new_size = snprintf(read_query, size, "%s WHERE %s %s %u", priv->read_query_start, dv[qrc->field_id].name, op, PTYPE_UINT32_GETVAL(qrc->value));
 						break;
 					case POSTGRES_PTYPE_UINT64:
-						new_size = snprintf(read_query, size, "%s WHERE %s %s %llu", priv->read_query_start, dv[qrc->field_id].name, op, PTYPE_UINT64_GETVAL(qrc->value));
+						new_size = snprintf(read_query, size, "%s WHERE %s %s %llu", priv->read_query_start, dv[qrc->field_id].name, op, (unsigned long long) PTYPE_UINT64_GETVAL(qrc->value));
 						break;
 					case POSTGRES_PTYPE_STRING:
 						new_size = snprintf(read_query, size, "%s WHERE %s %s '%s'", priv->read_query_start, dv[qrc->field_id].name, op, PTYPE_STRING_GETVAL(qrc->value));
